--- Server/src/main/java/JGS/CasperEvent/domain/event/dto/ResponseDto/rushEventResponseDto/LotteryEventWinnerListResponseDto.java
@@ -1,9 +1,0 @@
-package JGS.CasperEvent.domain.event.dto.ResponseDto.rushEventResponseDto;
-
-import JGS.CasperEvent.domain.event.dto.response.lottery.LotteryEventParticipantResponseDto;
-
-import java.util.List;
-
-public record LotteryEventWinnerListResponseDto(List<LotteryEventParticipantResponseDto> participantsList,
-                                                Boolean isLastPage, long totalParticipants) {
-}
