+++ Server/src/main/java/JGS/CasperEvent/domain/event/dto/response/ImageUrlResponseDto.java
@@ -1,4 +1,4 @@
-package JGS.CasperEvent.domain.event.dto.ResponseDto;
+package JGS.CasperEvent.domain.event.dto.response;
 
 public record ImageUrlResponseDto(String imageUrl) {
 }
