+++ Server/src/main/java/JGS/CasperEvent/domain/event/dto/response/ParticipantsListResponseDto.java
@@ -1,4 +1,4 @@
-package JGS.CasperEvent.domain.event.dto.ResponseDto;
+package JGS.CasperEvent.domain.event.dto.response;
 
 import java.util.List;
